Added a DrawLoop to allow for many syncronous animations
This commit fixes the problem where many animation frames would
be called together. It introduces a drawloop that exposes a
similar api to the animationframe and timeout api. This allowed
for a large speedup of the application as shown by the mouseChaser
which can now handle many more concurrent animations.

Changes
=======
- Wrote a draw loop
- Integrated the draw loop into the declarative
- Removed the closure since webpack already makes one
- Allowed the mouseChaser demo to use the draw loop