Simplified the drawloop code and updated demos
This commit uses a much simpler paradigm for the timeout code
that actually turns out to be more performant. We put the
timeouts in a queue and we always advance the queue. If the
timeout is not supposed to run yet, we force it to the end of
the queue. Since our queue is a linked list, this is an O(1)
operation and it is very fast.

Changes
=======
- Timeouts now use a queue
- updated the demos